Limace.InsulatingBasis: Insulating magnetic field basis

Limace.InsulatingBasis.inner_b0norm

Limace.InsulatingBasis.norm_B0fac!

Limace.InsulatingBasis.unitspherenorm

Limace.InsulatingBasis.inner_b0norm — Method

inner_b0norm(
    lmn_p,
    lmn_t
) -> SparseArrays.SparseMatrixCSC{ComplexF64, Int64}

Convenience functions to normalize a B₀ that is assembled from different components to have (4π/3)⁻¹ ∫₀¹ B⋅B dV = 1 Construct Aᵢⱼ = ∫₀¹ Bᵢ ⋅Bⱼ dV

Limace.InsulatingBasis.norm_B0fac! — Method

norm_B0fac!(B0fac, lmn_p, lmn_t)

Normalize B0fac (vector of scalars) corresponding to lmn_p and lmn_t basis elements. length(B0fac)=length(lmn_p)+length(lmn_t). Final array of B0fac will ensure that (4π/3)⁻¹ ∫₀¹ B⋅B dV = 1.

Limace.InsulatingBasis.unitspherenorm — Method

unitspherenorm(l, n) -> Any

The basis elements are normalized so that ∫₀^∞ B⋅B dV = 1. There is no external contribution for the toroidal components. For the poloidal components with n=1, there is an external contribution (r>1). To normalize the basis elements so that ∫₀¹ B⋅B dV = 1, the following norm function can be used: