Limace.C — Method

C(l, m) -> Any

Equation (113) in Ivers & Phillips (2008).

Limace._coriolis_ss — Method

_coriolis_ss(b::Basis, lmna, lmnb, r, wr; Ω) -> Any

Fallback for Coriolis term between poloidal and poloidal component, explicitly calculating the quadrature.

Limace._coriolis_st — Method

_coriolis_st(b::Basis, lmna, lmnb, r, wr; Ω) -> Any

Fallback for Coriolis term between poloidal and toroidal component, explicitly calculating the quadrature. Following eq. (114) in Ivers & Phillips (2008).

Limace._coriolis_ts — Method

_coriolis_ts(b::Basis, lmna, lmnb, r, wr; Ω) -> Any

Fallback for Coriolis term between toroidal and poloidal component, explicitly calculating the quadrature. Following eq. (112) in Ivers & Phillips (2008).

Limace._coriolis_tt — Method

_coriolis_tt(b::Basis, lmna, lmnb, r, wr; Ω) -> Any

Fallback for Coriolis term between toroidal and toroidal component, explicitly calculating the quadrature.