Limace._crossterm! — Method

_crossterm!(
    bi::Basis,
    B0::BasisElement{T0<:Basis, PT<:Helmholtz, T},
    bj::Basis,
    is,
    js,
    aijs,
    i0,
    j0,
    li,
    mi,
    lj,
    mj,
    rwrs,
    lmn2k_bi,
    lmn2k_bj,
    nrangefi,
    nrangefj,
    indf,
    EA;
    kwargs...
)

Fallback functions for _crossterm term for U0. Write specialized function to include e.g. bandedness in n.

Limace._crossterm! — Method

_crossterm!(
    bi::Basis,
    bj::Basis,
    B0::BasisElement{T0<:Basis, PT<:Helmholtz, T},
    is,
    js,
    aijs,
    i0,
    j0,
    li,
    mi,
    lj,
    mj,
    rwrs,
    lmn2k_bi,
    lmn2k_bj,
    nrangefi,
    nrangefj,
    indf,
    EA;
    kwargs...
)

Fallback functions for _crossterm! term for B0. Write specialized function to include e.g. bandedness in n.

Limace._induction_sSS — Method

_induction_sSS(
    ::Type{TA<:Basis},
    ::Type{TB<:Basis},
    ::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr;
    external
) -> Any

Computes radial integral of ∫Sᵢ⋅∇×(pⱼ×Sₖ) dV. Surface integral is given by Adam-Gaunt variable.

Limace._induction_sST — Method

_induction_sST(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Tᵢ⋅∇×(pⱼ×Sₖ) dV. Surface integral is given by Elsasser variable.

Limace._induction_sTS — Method

_induction_sTS(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Sᵢ⋅∇×(pⱼ×Tₖ) dV. Surface integral is given by Elsasser variable.

Limace._induction_sTT — Method

_induction_sTT(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Tᵢ⋅∇×(pⱼ×Tₖ) dV. Surface integral is given by Adam-Gaunt variable.

Limace._induction_tSS — Method

_induction_tSS(
    ::Type{TA<:Basis},
    ::Type{TB<:Basis},
    ::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr;
    external
) -> Any

Computes radial integral of ∫Sᵢ⋅∇×(qⱼ×Sₖ) dV. Surface integral is given by Elsasser variable.

Limace._induction_tST — Method

_induction_tST(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Tᵢ⋅∇×(qⱼ×Sₖ) dV. Surface integral is given by Adam-Gaunt variable.

Limace._induction_tTT — Method

_induction_tTT(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Tᵢ⋅∇×(qⱼ×Tₖ) dV. Surface integral is given by Adam-Gaunt variable.

Limace.adamgaunt_ljs — Method

adamgaunt_ljs(li, l0, mj, lmax) -> Any

For a combination of li, l0 and lj only the even lj that satisfy |l0-li|≤lj≤l0+li are nonzero for the Adam-Gaunt variables.

Limace.adamgaunt_mjs — Method

adamgaunt_mjs(mi, m0) -> Any

For a combination of mi, m0 and mj only mj=mi-m0 is nonzero for the Adam-Gaunt variables.

Limace.elsasser_ljs — Method

elsasser_ljs(li, l0, mj, lmax) -> Any

For a combination of li, l0 and lj only the odd lj that satisfy |l0-li|≤lj≤l0+li are nonzero for the Elsasser variables.

Limace.elsasser_mjs — Method

elsasser_mjs(mi, m0) -> Any

For a combination of mi, m0 and mj only mj=mi-m0 is nonzero for the Elsasser variables.

Limace.induction — Method

induction(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Poloidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a poloidal background velocity U0, a magnetic field basis bbi and a magnetic field basis bbj.

Limace.induction — Method

induction(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Toroidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a toroidal background velocity U0, a magnetic field basis bbi and a magnetic field basis bbj.

Limace.induction — Method

induction(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a poloidal background magnetic field B0, a magnetic field basis bbi and a velocity basis buj.

Limace.induction — Method

induction(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a toroidal background magnetic field B0, a magnetic field basis bbi and a velocity basis buj.

Limace.induction_threaded — Method

induction_threaded(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Poloidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Threaded version of induction

Limace.induction_threaded — Method

induction_threaded(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Toroidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Threaded version of induction

Limace.induction_threaded — Method

induction_threaded(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Threaded version of induction

Limace.induction_threaded — Method

induction_threaded(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Threaded version of induction