Limace.lorentz — Method

lorentz(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a poloidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.

Limace.lorentz — Method

lorentz(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a toroidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.

Limace.lorentz_threaded — Method

lorentz_threaded(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a poloidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.

Limace.lorentz_threaded — Method

lorentz_threaded(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a toroidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.