Poly

Polynomial functions and conveniences.

Polynomials

Limace.Poly.jacobi — Function

jacobi(n, a, b, x) -> Any

Jacobi polynomial

math J_n^{(a,b)}(x)`

Limace.Poly.ylm — Function

ylm(ℓ::Int64, m::Int64, θ, φ) -> Any

Spherical harmonic $Y_l^m$ in full norm, i.e.

\[\int Y_l^mY_i^j\, \sin(\theta)\,\mathrm{d}\theta\mathrm{d}\phi = \delta_{li}\delta_{mj}\]

where $\theta$ is the colatitude and $\phi$ the azimuthal angle.

Integral functions

Limace.Poly.adamgaunt — Function

adamgaunt(la,lb,lc,ma,mb,mc)

Adam-Gaunt integral (James, 1973)

\[A_{abc} = \oint\int Y_aY_bY_c\sin\theta\,\mathrm{d}\theta\mathrm{d}\phi\]

Limace.Poly.elsasser — Function

elsasser(la,lb,lc,ma,mb,mc)

Elsasser integral (James, 1973)

\[E_{abc} = \oint\int Y_c\left( \frac{\partial Y_a}{\partial \theta} \frac{\partial Y_b}{\partial \phi} - \frac{\partial Y_a}{\partial \phi}\frac{\partial Y_b}{\partial \theta} \right)\,\mathrm{d}\theta\mathrm{d}\phi\]

Limace.Poly.inners — Function

inners(s, s2, l, r) -> Any

\[\frac{l(l+1)}{r^2}\left( l(l+1)s s_2 + \frac{\partial r s}{\partial r}\frac{\partial r s_2}{\partial r}\right)\]

Radial function to be integrated in radius when computing the inner product of two poloidal vectors.

Limace.Poly.innert — Function

innert(t, t2, l::Int64, r) -> Any

\[l(l+1)t t_2\]

Radial function to be integrated in radius when computing the inner product of two toroidal vectors.

Miscellaneous functions

Limace.Poly.p — Function

p(l) -> Any

$p(l) = l(l+1)$ following notation of Ivers and Phillips (2008)

Limace.Poly._∂ll — Function

_∂ll(f, l, l1, r) -> Any

\[\partial_l^{l_1} = \begin{cases} \frac{\partial f}{\partial r} + \frac{l+1}{r}f \quad \mathrm{if}\, l_1 = l-1\\ \frac{\partial f}{\partial r} - \frac{l}{r}f \quad \mathrm{if}\, l_1 = l+1 \end{cases}\]

Equation (25) in Ivers and Phillips (2008).

Limace.Poly.D — Function

D(f, l, r) -> Any

\[D_l(f) = \frac{\partial^2 f}{\partial r^2} +\frac{2}{r}\frac{\partial f}{\partial r} - \frac{l(l+1)}{r^2}f\]

Below equation (31) in Ivers and Phillips (2008).

Limace.Poly.dylmdθ — Function

dylmdθ(l, m, θ, ϕ) -> Any

Derivative of spherical harmonic $Y_l^m$ in $\theta$.

Limace.Poly.dylmdϕ — Function

dylmdϕ(l, m, θ, ϕ) -> Any

Derivative of spherical harmonic $Y_l^m$ in $\phi$.