Postprocessing

A non-exhaustive collection of spectral to spatial discretization and postprocessing routines.

Spatial discretization

In the Limace.Discretization submodule, some methods to discretize BasisElement and eigenvectors are provided.

Limace.Discretization.discretize — Method

discretize(b::BasisElement{TB,TP,T}, r, θ, ϕ, V::Volume=Sphere()) where {TB<:Basis,TP<:Helmholtz,T<:Number}

Discretize a basis element b at the given coordinates (r, θ, ϕ) in the volume V.

Limace.Discretization.spectospat — Method

spectospat(
    coeffs::Array{T<:ComplexF64, 1},
    u::Basis,
    b::Basis,
    r::Union{Float64, AbstractVector{Float64}},
    nθ::Int64,
    nϕ::Int64;
    kwargs...
) -> Tuple{Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Vector{Float64}, StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}, Int64}}

Transform eigenvector containing spectral coefficients to two-dimensional vector fields at 1 x nθ x nϕ grid points at radius r. Returns ur,uθ,uϕ, br,bθ,bϕ, θ,ϕ.

Limace.Discretization.spectospat — Method

spectospat(
    coeffs::Array{T<:ComplexF64, 1},
    u::Basis,
    b::Basis,
    nr::Int64,
    nθ::Int64,
    nϕ::Int64;
    kwargs...
) -> Tuple{Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Any, Vector{Float64}, StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}, Int64}}

Transform eigenvector containing spectral coefficients to three-dimensional vector fields at nr x nθ x nϕ grid points at radius r. Returns ur,uθ,uϕ, br,bθ,bϕ, r,θ,ϕ.

Limace.Discretization.spectospat — Method

spectospat(
    coeffs::Array{T<:ComplexF64, 1},
    u::Basis,
    r::Float64,
    nθ::Int64,
    nϕ::Int64;
    kwargs...
) -> Tuple{Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Vector{Float64}, StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}, Int64}}

Transform eigenvector containing spectral coefficients to two-dimensional vector field at 1 x nθ x nϕ grid points at radius r. Returns ur,uθ,uϕ, θ,ϕ.

Limace.Discretization.spectospat — Method

spectospat(
    coeffs::Array{T<:ComplexF64, 1},
    u::Basis,
    nr::Int64,
    nθ::Int64,
    nϕ::Int64;
    kwargs...
) -> Tuple{Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Any, Vector{Float64}, StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}, Int64}}

Transform eigenvector containing spectral coefficients to three-dimensional vector field at nr x nθ x nϕ grid points. Returns ur,uθ,uϕ, r,θ,ϕ.

Filtering of spectrum

Limace.Processing.eigenvalue_filter — Function

eigenvalue_filter(evals1, evals2; λtol)

Find all indices of evals1 and evals2 for which findall(y->any(x->isapprox(x,y; rtol=λtol), evals2),evals1). Multithreaded.

Limace.Processing.numerical_filter — Function

numerical_filter(
    evecs1,
    evecs2,
    evals1,
    evals2,
    u1,
    u2,
    b1,
    b2;
    threshc,
    λtol,
    threads,
    epeakratiothresh,
    kwargs...
)

Find numerically converged eigensolutions between two resolutions.

Limace.Processing.observability_filter — Function

observability_filter(
    evals,
    evecs,
    u,
    b;
    ωlow,
    ωhigh,
    Qlow,
    lthresh
)

Find observationally relevant solutions, based on the poloidal magnetic field at the surface. The function takes the arguments evals, evecs, u, and b which are the eigenvalues, eigenvectors, velocity basis, and magnetic field basis, respectively. Keyword arguments ωlow, ωhigh, Qlow, and lthresh are used to filter the eigenvalues based on their frequency, quality factor, and the maximum allowed peak degree in energy.

Energy and spectra

Limace.Processing.energies — Function

energies(problem::LimaceProblem)

Compute the energies for all bases problem.bases of all eigenvectors problem.sol.vectors if problem.solved==true.

Limace.Processing.epeak_etrunc_cartesian — Function

epeak_etrunc_cartesian(evecs, u, b)

Compute the ratio of peak energy to energy at truncation degree (max between two last Cartesian degrees) in toroidal/poloidal kinetic/magnetic energy. evecs are the eigenvectors computed using the velocity basis u and magnetic field basis b.

Limace.Processing.lmn_n — Function

lmn_n(u, b)

Get all `(l,m,n)` that correspond to the poloidal and toroidal component of `u` and `b` basis at each Cartesian
degree `ñ ∈ 1:N`.

Limace.Processing.spectrum — Function

spectrum(evecs, u, b; lmn)

Calculate spectrum for all eigenvectors evecs comprised of velocity basis u and magnetic field basis b. Use keyword lmn=1,2,3 to select l=1, m=2 or n=3.

spectrum(problem; lmn)

Calculate spectrum for all eigenvectors problem.sol.vectors, comprised of velocity basis u (problem.bases[1]!) and magnetic field basis b (problem.bases[2]!). Use keyword lmn=1,2,3 to compute spectra in $l$ (lmn=1), $m$ (lmn=2) or in $n$ (lmn=3).

Limace.Processing.spectrum_cartesian — Function

spectrum_cartesian(evecs, u, b)

Compute poloidal/toroidal kinetic/magnetic spectra of evecs as a function of max. Cartesian monomial degree. evecs are the eigenvectors computed using the velocity basis u and magnetic field basis b.