Limace.Processing

A non-exhaustive collection of postprocessing routines.

Limace.Processing.eigenvalue_filter

Limace.Processing.epeak_etrunc_cartesian

Limace.Processing.lmn_n

Limace.Processing.spectrum

Limace.Processing.spectrum_cartesian

Limace.Processing.eigenvalue_filter — Method

eigenvalue_filter(evals1, evals2; λtol=1e-3)

Find all indices of `evals1` and `evals2` for which `findall(y->any(x->isapprox(x,y; rtol=λtol), evals2),evals1)`.
Multithreaded.

Limace.Processing.epeak_etrunc_cartesian — Method

epeaketrunccartesian(evecs, u, b)

Compute the ratio of peak energy to energy at truncation degree
(max between two last Cartesian degrees) in toroidal/poloidal kinetic/magnetic energy.

Limace.Processing.lmn_n — Method

lmn_n(u,b)

Get all `(l,m,n)` that correspond to the poloidal and toroidal component of `u` and `b` basis at each Cartesian
degree `ñ ∈ 1:N`.

Limace.Processing.spectrum — Method

spectrum(evecs, u, b; lmn=1)

calculate spectrum for all eigenvectors `evecs` comprised of velocity basis `u` and magnetic field basis `b`. 
Use keyword `lmn=1,2,3` to select `l=1`, `m=2` or `n=3`.

Limace.Processing.spectrum_cartesian — Method

spectrum_cartesian(evecs, u, b)

Compute poloidal/toroidal kinetic/magnetic spectra of `evecs` as a function of max. Cartesian monomial degree.