Bases

Several bases are readily implemented for full sphere geometries and satisfying certain boundary and orthogonality conditions. The most commonly encountered ones are outlined here and a small introduction into how to write a new basis is given at the end.

Inviscid velocity basis

For an inviscid fluid, we only require $\mathbf{u}\cdot\mathbf{n}=0$ at the boundary. This is known as the non-penetration condition. For poloidal and toroidal decomposition, this boils down to the requirement that the poloidal scalar vanishes at the boundary. The toroidal scalar remains unconstrained and we have free slip.

Limace.Bases.s — Method

s(_::Type{Basis{Inviscid}}, V::Volume, l, m, n, r) -> Any

\[s_{l,n,m}(r) = f_{l,n}(1-r^2)r^l J_n^{(1,l+1/2)}(2r^2-1)\]

with

\[f_{l,n} = \sqrt{\frac{5+2l+4n}{4l(l+1)(n+1)^2}}\]

Livermore (2014) (5.6), normalized to unit energy.

Limace.Bases.t — Method

t(_::Type{Basis{Inviscid}}, V::Volume, l, m, n, r) -> Any

\[t_{l,n,m}(r) = f_{l,n}r^l J_n^{(0,l+1/2)}(2r^2-1)\]

with

\[f_{l,n} = \sqrt{\frac{3+2l+4n}{l(l+1)}}\]

Livermore (2014) (5.1), normalized to unit energy.

Viscous velocity basis

For a viscous fluid, in addition to the non-penetration condition, we require no-slip, so that in total $\mathbf{u} = \mathbf{0}$ at the boundary. Chen et al. (2018) have written a basis that satisfies these conditions, and requires orthogonality w.r.t the vector Laplacian, so that

\[\int \mathbf{u}_i^*\cdot\boldsymbol{\nabla}^2\mathbf{u}_j\,\mathrm{d}V = \delta_{ij}.\]

This is desirable, as the resulting projections of all considered forces are banded (if only combined with another basis that satisfies the appropriate orthogonality).

Limace.Bases.s — Method

s(_::Type{Basis{Viscous}}, V::Volume, l, m, n, r) -> Any

\[s_{l,m,n} = f_{l,n}r^l\left( (2l+4n+1)J_{n+1}^{(0,l+1/2)}(2r^2-1)-2(2l+4n+3)J_{n}^{(0,l+1/2)}(2r^2-1) + (2l+4n+5)J_{n-1}^{(0,l+1/2)}(2r^2-1) \right)\]

with

\[f_{l,n} = \left( 2l(l+1)(2l+4n+1)(2l+4n+3)(2l+4n+5) \right)^{-1/2}\]

Chen et al. (2018) (2.38), (2.39) poloidal scalar, orthogonal w.r.t ∫ u⋅∇²u dV with 0 ≤ r ≤ 1.

Limace.Bases.t — Method

t(_::Type{Basis{Viscous}}, V::Volume, l, m, n, r) -> Any

\[t_{l,m,n}(r) = f_{l,n}r^l\left( J_n^{(0,l+1/2)}(2r^2-1) - J_{n-1}^{(0,l+1/2)}(2r^2-1)\right)\]

with

\[f_{l,n}=\left(l(l+1)/(2l+4n-1)+1/(2l+4n+3)\right)^{-1/2}\]

Chen et al. (2018) toroidal scalar, orthogonal w.r.t ∫ u⋅∇²u dV with 0 ≤ r ≤ 1.

Insulating magnetic field basis

Limace.Bases.s — Method

s(_::Type{Basis{Insulating}}, V::Volume, l, m, n, r) -> Any

\[s_{l,m,n}(r) = f_{l,n}r^l\left( (2l+4n+3)J_{n}^{(0,l+1/2)}(2r^2-1) -2(2l+4n-1)J_{n-1}^{(0,l+1/2)}(2r^2-1) + (2l+4n+1)J_{n-2}^{(0,l+1/2)}(2r^2-1) \right)\]

with

\[f_{l,n} = \left( 2l(l+1)(2l+4n-3)(2l+4n-1)(2l+4n+1)\right)^{-1/2}\]

Gerick and Livermore (2024) (A6)

Limace.Bases.t — Method

t(_::Type{Basis{Insulating}}, V::Volume, l, m, n, r) -> Any

\[t_{l,m,n}(r) = f_{l,n}r^l\left( J_{n}^{(0,l+1/2)}(2r^2-1) - J_{n-1}^{(0,l+1/2)}(2r^2-1)\right)\]

with

\[f_{l,n} = \left( l(l+1)/(2l+4n-1) + 1/(2l+4n+3) \right)^{-1/2}\]

Gerick and Livermore (2024) (A7)

Limace.InsulatingBasis.inner_b0norm — Method

inner_b0norm(
    lmn_p,
    lmn_t
) -> SparseArrays.SparseMatrixCSC{ComplexF64, Int64}

Convenience functions to normalize a B₀ that is assembled from different components to have (4π/3)⁻¹ ∫₀¹ B⋅B dV = 1 Construct Aᵢⱼ = ∫₀¹ Bᵢ ⋅Bⱼ dV

Limace.InsulatingBasis.norm_B0fac! — Method

norm_B0fac!(B0fac, lmn_p, lmn_t)

Normalize B0fac (vector of scalars) corresponding to lmn_p and lmn_t basis elements. length(B0fac)=length(lmn_p)+length(lmn_t). Final array of B0fac will ensure that (4π/3)⁻¹ ∫₀¹ B⋅B dV = 1.

Limace.InsulatingBasis.unitspherenorm — Method

unitspherenorm(l, n) -> Any

The basis elements are normalized so that ∫₀^∞ B⋅B dV = 1. There is no external contribution for the toroidal components. For the poloidal components with n=1, there is an external contribution (r>1). To normalize the basis elements so that ∫₀¹ B⋅B dV = 1, the following norm function can be used:

Unconstrained

Limace.Bases.s — Method

s(
    _::Type{Basis{Limace.UnconstrainedBasis.Unconstrained}},
    V::Volume,
    l,
    m,
    n,
    r
) -> Any

\[s_{l,n,m}(r) = f_{l,n}r^l\left(J_n^{(0,l+1/2)}(2r^2-1) - J_{n-1}^{(0,l+1/2)}(2r^2-1) \right)\]

with

\[f_{l,n} = \left( l(l+1)(2l+4n+1)\right)^{-1/2}\]

Livermore (2014) (5.3), normalized to unit energy ∫u⋅u dV = 1.

Limace.Bases.t — Method

t(
    _::Type{Basis{Limace.UnconstrainedBasis.Unconstrained}},
    V::Volume,
    l,
    m,
    n,
    r
) -> Any

\[t_{l,n,m}(r) = f_{l,n}r^l J_n^{(0,l+1/2)}(2r^2-1)\]

with

\[f_{l,n} = \sqrt{\frac{3+2l+4n}{l(l+1)}}\]

Livermore (2014) (5.1), normalized to unit energy ∫u⋅u dV = 1.

Implementing a new basis

All bases are defined in a submodule, as parametric types T of a Basis{T} basis struct and associated with a BoundaryCondition.

Limace.Bases.Basis — Type

struct Basis{T}

N::Int64

m::UnitRange{Int64}

n::UnitRange{Int64}

BC::BoundaryCondition

V::Volume

params::Dict{Symbol, Float64}

Limace.Bases.BasisElement — Type

struct BasisElement{TB<:Basis, PT<:Helmholtz, T<:Number}

lmn::Tuple{Int64, Int64, Int64}

factor::Number

Limace.Bases.BoundaryCondition — Type

abstract type BoundaryCondition

Limace.Bases.InsulatingBC — Type

struct InsulatingBC <: BoundaryCondition

Limace.Bases.InviscidBC — Type

struct InviscidBC <: BoundaryCondition

Limace.Bases.NoBC — Type

struct NoBC <: BoundaryCondition

Limace.Bases.NoSlipBC — Type

struct NoSlipBC <: BoundaryCondition

Limace.Bases.PerfectlyConductingBC — Type

struct PerfectlyConductingBC <: BoundaryCondition

We can follow the example of Limace.InsulatingBasisNoBC, which implements an insulating magnetic field basis, with the boundary condition explicitly imposed afterwards.

Define a new module in a .jl file in the bases folder, and import several components of Limace.jl that are needed:

module InsulatingBasisNoBC

using SparseArrays

using SparseArrays
using LinearAlgebra
using DocStringExtensions

using ..Bases
using ..UnconstrainedBasis, ..InviscidBasis, ..InsulatingBasis
using ..Utils
using ..Poly

using ..Bases: nrange_p, nrange_t, nrange_p_bc, nrange_t_bc, np, nt, t, s, bcs_p, bcs_t, lmn_p_l, lmn_t_l, lmn_p, lmn_t, lmn2k_p_dict, lmn2k_t_dict, lpmax, ltmax, Sphere
import ..Bases: lpmax, ltmax, lmn_t, lmn_p, _nrange_p, _nrange_t, np, nt, t, s, nrange_p_bc, nrange_t_bc, bcs_p, bcs_t
using ..Poly: ∂

We define our new basis, with the right boundary condition. Here, we choose BC=NoBC(), as we impose the boundary condition explicitly afterwards.

export InsulatingNoBC

struct InsulatingNoBC end

InsulatingNoBC(N; kwargs...) = Basis{InsulatingNoBC}(; N, BC=NoBC(), V=Sphere(), kwargs...)

We need to define a method for s and t for our basis. In this example, we just use the Unconstrained basis.

s(::Type{Basis{InsulatingNoBC}}, V::Volume, l,m,n,r)  = s(Basis{Unconstrained}, V, l,m,n,r) 
t(::Type{Basis{InsulatingNoBC}}, V::Volume, l,m,n,r)  = t(Basis{Unconstrained}, V, l,m,n,r) 

A few additional functions need to be defined to determine the radial and spherical harmonic degrees.

@inline _nrange_p(b::Basis{InsulatingNoBC}, l) = 0:((b.N-l+1)÷2)
@inline _nrange_t(b::Basis{InsulatingNoBC}, l) = 0:((b.N-l)÷2)

@inline lpmax(b::Basis{InsulatingNoBC}) = b.N
@inline ltmax(b::Basis{InsulatingNoBC}) = b.N

If we have no basis imposed in the poloidal and toroidal scalars (NoBC()), we need to define the explicit boundary condition functions for the toroidal (bcs_t) and poloidal (bcs_p) scalars.

@inline function bcs_t(b::Basis{InsulatingNoBC})
    fs = (@inline((l, n) -> t(Basis{InsulatingNoBC}, b.V, l, 0, n, 1.0)),)
    return fs
end

@inline function bcs_p(b::Basis{InsulatingNoBC})
    fs = (@inline((l, n) -> ∂(r -> s(Basis{InsulatingNoBC}, b.V, l, 0, n, r), 1.0) + (l + 1) * s(Basis{InsulatingNoBC}, b.V, l, 0, n, 1.0)),)
    return fs
end

end #module

Here, these are given as evaluations at the surface, r=1.0. We use automatic differentiation to compute the derivatives in r using ForwardDiff.jl.