Forces

Different forces included in Limace.jl.

So far, there is no heat equation in Limace.jl, so the forces include only the mass-term (or Limace.inertial), the Coriolis force (Limace.coriolis), the Lorentz force (Limace.lorentz) and diffusion (Limace.diffusion). In the induction equation we have the magnetic induction term $\nabla\times\mathbf{u}\times\mathbf{B}$ (Limace.induction). Some details for each of these forcings is given on the respective pages.

Inertial

\[\int \mathbf{u}_i^* \cdot \mathbf{u}_j\,\mathrm{d}V,\]

Limace._inertial_ss — Method

_inertial_ss(b::Basis, lmna, lmnb, r, wr; external) -> Any

Limace._inertial_tt — Method

_inertial_tt(b::Basis, lmna, lmnb, r, wr) -> Any

Limace.inertial — Method

inertial(
    b::Basis;
    ...
) -> LinearAlgebra.UniformScaling{Float64}
inertial(b::Basis, ::Type{T<:Number}; external) -> Any

Compute the Galerkin projection matrix of basis b onto itself, i.e. the inner products. Also known as the mass matrix.

Coriolis

The necessary functions to compute Galerkin projections on the Coriolis term

\[\int \mathbf{u}_i^* \cdot 2\mathbf{e}_z\times\mathbf{u}_j\,\mathrm{d}V,\]

where $\mathbf{u}_{i,j}$ are either poloidal or toroidal basis vectors.

Limace.C — Method

C(l, m) -> Any

Equation (113) in Ivers and Phillips (2008).

Limace._coriolis_ss — Method

_coriolis_ss(b::Basis, lmna, lmnb, r, wr; Ω) -> Any

Fallback for Coriolis term between poloidal and poloidal component, explicitly calculating the quadrature.

Limace._coriolis_st — Method

_coriolis_st(b::Basis, lmna, lmnb, r, wr; Ω) -> Any

Fallback for Coriolis term between poloidal and toroidal component, explicitly calculating the quadrature. Following eq. (114) in Ivers and Phillips (2008).

Limace._coriolis_ts — Method

_coriolis_ts(b::Basis, lmna, lmnb, r, wr; Ω) -> Any

Fallback for Coriolis term between toroidal and poloidal component, explicitly calculating the quadrature. Following eq. (112) in Ivers and Phillips (2008).

Limace._coriolis_tt — Method

_coriolis_tt(b::Basis, lmna, lmnb, r, wr; Ω) -> Any

Fallback for Coriolis term between toroidal and toroidal component, explicitly calculating the quadrature.

Limace.coriolis — Method

coriolis(
    b;
    Ω
) -> SparseArrays.SparseMatrixCSC{ComplexF64, Int64}

Compute the sparse Galerkin projection matrix, by projecting the basis b onto the Coriolis operator.

Diffusion

\[\int \mathbf{u}_i^* \cdot \boldsymbol{\nabla}^2\mathbf{u}_j\,\mathrm{d}V\]

Limace._diffusion_ss — Method

_diffusion_ss(b::Basis, lmna, lmnb, r, wr; external) -> Any

Diffusion term of the poloidal components. Set external=true to include the contribution of an external poloidal magnetic field matching at the surface.

Limace._diffusion_tt — Method

_diffusion_tt(b::Basis, lmna, lmnb, r, wr) -> Any

Diffusion term of the toroidal components.

Limace.diffusion — Method

diffusion(b::Number; ...)
diffusion(
    b::Basis,
    ::Type{T<:Number};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv<:(Complex{T} where T<:Number)

Compute the Galerkin projection matrix of the basis b onto the vector Laplacian. When keyword external=true, the integral is computed over all space, assuming continuity of the poloidal field and a scalar potential in the exterior domain.

Limace.diffusion_threaded — Method

diffusion_threaded(b::Number; ...)
diffusion_threaded(
    b::Basis,
    ::Type{T<:Number};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Induction

\[\int \mathbf{b}_i^* \cdot \boldsymbol{\nabla}\times\left(\mathbf{u}_j\times\mathbf{B}_k\right)\,\mathrm{d}V\]

Limace._crossterm! — Method

_crossterm!(
    bi::Basis,
    B0::BasisElement{T0<:Basis, PT<:Helmholtz, T},
    bj::Basis,
    is,
    js,
    aijs,
    i0,
    j0,
    li,
    mi,
    lj,
    mj,
    rwrs,
    lmn2k_bi,
    lmn2k_bj,
    nrangefi,
    nrangefj,
    indf,
    EA;
    kwargs...
)

Fallback functions for _crossterm term for U0. Write specialized function to include e.g. bandedness in n.

Limace._crossterm! — Method

_crossterm!(
    bi::Basis,
    bj::Basis,
    B0::BasisElement{T0<:Basis, PT<:Helmholtz, T},
    is,
    js,
    aijs,
    i0,
    j0,
    li,
    mi,
    lj,
    mj,
    rwrs,
    lmn2k_bi,
    lmn2k_bj,
    nrangefi,
    nrangefj,
    indf,
    EA;
    kwargs...
)

Fallback functions for _crossterm! term for B0. Write specialized function to include e.g. bandedness in n.

Limace._induction_sSS — Method

_induction_sSS(
    ::Type{TA<:Basis},
    ::Type{TB<:Basis},
    ::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr;
    external
) -> Any

Computes radial integral of ∫Sᵢ⋅∇×(pⱼ×Sₖ) dV. Surface integral is given by Adam-Gaunt variable.

Limace._induction_sST — Method

_induction_sST(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Tᵢ⋅∇×(pⱼ×Sₖ) dV. Surface integral is given by Elsasser variable.

Limace._induction_sTS — Method

_induction_sTS(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Sᵢ⋅∇×(pⱼ×Tₖ) dV. Surface integral is given by Elsasser variable.

Limace._induction_sTT — Method

_induction_sTT(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Tᵢ⋅∇×(pⱼ×Tₖ) dV. Surface integral is given by Adam-Gaunt variable.

Limace._induction_tSS — Method

_induction_tSS(
    ::Type{TA<:Basis},
    ::Type{TB<:Basis},
    ::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr;
    external
) -> Any

Computes radial integral of ∫Sᵢ⋅∇×(qⱼ×Sₖ) dV. Surface integral is given by Elsasser variable.

Limace._induction_tST — Method

_induction_tST(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Tᵢ⋅∇×(qⱼ×Sₖ) dV. Surface integral is given by Adam-Gaunt variable.

Limace._induction_tTT — Method

_induction_tTT(
    _::Type{TA<:Basis},
    _::Type{TB<:Basis},
    _::Type{TC<:Basis},
    V::Volume,
    lmna,
    lmnb,
    lmnc,
    r,
    wr
) -> Any

Computes radial integral of ∫Tᵢ⋅∇×(qⱼ×Tₖ) dV. Surface integral is given by Adam-Gaunt variable.

Limace.adamgaunt_ljs — Method

adamgaunt_ljs(li, l0, mj, lmax) -> Any

For a combination of li, l0 and lj only the even lj that satisfy |l0-li|≤lj≤l0+li are nonzero for the Adam-Gaunt variables.

Limace.adamgaunt_mjs — Method

adamgaunt_mjs(mi, m0) -> Any

For a combination of mi, m0 and mj only mj=mi-m0 is nonzero for the Adam-Gaunt variables.

Limace.elsasser_ljs — Method

elsasser_ljs(li, l0, mj, lmax) -> Any

For a combination of li, l0 and lj only the odd lj that satisfy |l0-li|≤lj≤l0+li are nonzero for the Elsasser variables.

Limace.elsasser_mjs — Method

elsasser_mjs(mi, m0) -> Any

For a combination of mi, m0 and mj only mj=mi-m0 is nonzero for the Elsasser variables.

Limace.induction — Method

induction(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Poloidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a poloidal background velocity U0, a magnetic field basis bbi and a magnetic field basis bbj.

Limace.induction — Method

induction(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Toroidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a toroidal background velocity U0, a magnetic field basis bbi and a magnetic field basis bbj.

Limace.induction — Method

induction(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a poloidal background magnetic field B0, a magnetic field basis bbi and a velocity basis buj.

Limace.induction — Method

induction(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a toroidal background magnetic field B0, a magnetic field basis bbi and a velocity basis buj.

Limace.induction_threaded — Method

induction_threaded(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Poloidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Threaded version of induction

Limace.induction_threaded — Method

induction_threaded(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Toroidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Threaded version of induction

Limace.induction_threaded — Method

induction_threaded(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Threaded version of induction

Limace.induction_threaded — Method

induction_threaded(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Threaded version of induction

Lorentz

\[\int \mathbf{b}_i^* \cdot \left(\boldsymbol{\nabla}\times\mathbf{b}_j\times\mathbf{b}_k\right)\,\mathrm{d}V\]

Limace.lorentz — Method

lorentz(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a poloidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.

Limace.lorentz — Method

lorentz(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a toroidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.

Limace.lorentz_threaded — Method

lorentz_threaded(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a poloidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.

Limace.lorentz_threaded — Method

lorentz_threaded(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a toroidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.

Explicit boundary condition operator

Limace.boundarycondition! — Method

boundarycondition!(A, b::Basis)
boundarycondition!(A, b::Basis, ::Type{T<:Number})

Limace.boundarycondition — Method

boundarycondition(b::Number)
boundarycondition(
    b::Basis,
    ::Type{T<:Number}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv<:(Complex{T} where T<:Number)

Limace.zero_boundarycondition! — Method

zero_boundarycondition!(A, b::Basis)
zero_boundarycondition!(A, b::Basis, ::Type{T<:Number})