Forces

Different forces included in Limace.jl.

So far, there is no heat equation in Limace.jl, so the forces include only the mass-term (or Limace.inertial), the Coriolis force (Limace.coriolis), the Lorentz force (Limace.lorentz) and diffusion (Limace.diffusion). In the induction equation we have the magnetic induction term $\nabla\times\mathbf{u}\times\mathbf{B}$ (Limace.induction). Some details for each of these forcings is given on the respective pages.

Inertial

\[\int \mathbf{u}_i^* \cdot \mathbf{u}_j\,\mathrm{d}V,\]

Limace.inertial — Function

inertial(
    b::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{ComplexF64, Int64}

Compute the Galerkin projection matrix of basis b onto itself, i.e. the inner products. Also known as the mass matrix.

Coriolis

The necessary functions to compute Galerkin projections on the Coriolis term

\[\int \mathbf{u}_i^* \cdot 2\mathbf{e}_z\times\mathbf{u}_j\,\mathrm{d}V,\]

where $\mathbf{u}_{i,j}$ are either poloidal or toroidal basis vectors.

Limace.coriolis — Function

coriolis(
    b;
    Ω
) -> SparseArrays.SparseMatrixCSC{ComplexF64, Int64}

Compute the sparse Galerkin projection matrix, by projecting the basis b onto the Coriolis operator.

Diffusion

\[\int \mathbf{u}_i^* \cdot \boldsymbol{\nabla}^2\mathbf{u}_j\,\mathrm{d}V\]

Limace.diffusion — Function

diffusion(
    b::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

Compute the Galerkin projection matrix of the basis b onto the vector Laplacian. When keyword external=true, the integral is computed over all space, assuming continuity of the poloidal field and a scalar potential in the exterior domain.

Induction

\[\int \mathbf{b}_i^* \cdot \boldsymbol{\nabla}\times\left(\mathbf{u}_j\times\mathbf{B}_k\right)\,\mathrm{d}V\]

Limace.induction — Function

induction(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a poloidal background magnetic field B0, a magnetic field basis bbi and a velocity basis buj.

induction(
    bbi::Basis,
    buj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T};
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a toroidal background magnetic field B0, a magnetic field basis bbi and a velocity basis buj.

induction(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Poloidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a poloidal background velocity U0, a magnetic field basis bbi and a magnetic field basis bbj.

induction(
    bbi::Basis,
    U0::BasisElement{T0<:Basis, Toroidal, T},
    bbj::Basis;
    external
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the induction term for a toroidal background velocity U0, a magnetic field basis bbi and a magnetic field basis bbj.

Lorentz

\[\int \mathbf{u}_i^* \cdot \left(\boldsymbol{\nabla}\times\mathbf{b}_j\times\mathbf{b}_k\right)\,\mathrm{d}V\]

Limace.lorentz — Function

lorentz(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Poloidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a poloidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.

lorentz(
    bui::Basis,
    bbj::Basis,
    B0::BasisElement{T0<:Basis, Toroidal, T}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv

Computes the Lorentz term for a toroidal background magnetic field B0, a velocity basis bui and a magnetic field basis bbj.

Explicit boundary condition operator

Limace.boundarycondition! — Method

boundarycondition!(A, b::Basis)
boundarycondition!(A, b::Basis, ::Type{T<:Number})

Limace.boundarycondition — Method

boundarycondition(b::Number)
boundarycondition(
    b::Basis,
    ::Type{T<:Number}
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv<:(Complex{T} where T<:Number)

Limace.zero_boundarycondition! — Method

zero_boundarycondition!(A, b::Basis)
zero_boundarycondition!(A, b::Basis, ::Type{T<:Number})